Controllers were implemented inside the modules. Modifications in routes. async / await implemented in controllers.
