+++ server/controllers/aboutController.js
@@ -1,0 +1,5 @@
+exports.view = (req, res) => {
+  res.render('about', {
+    page: 'Sobre Nosotros',
+  });
+}
